Self-Healing-Bot/requirements.txt
kubernetes
prometheus-client
opentelemetry-sdk
pytest
python-dotenv
aiohttp